Refactor chatbot module by extracting selection handlers
Moved complex `extend_selection` and `edit_selection` handlers from `plugin/modules/chatbot/__init__.py` to a dedicated `selection.py` file to reduce the complexity of the `ChatbotModule` class, keeping it focused on initialization and dispatching.
